perf(router) cache service loading in router
The current implementation of the router loads everything from the
database, all the time.

Since it iterates over routes, and then it fetches the relevant
service for each route, this means that given a single service with
1000 routes, it will make 1000 requests for the database, returning
the same service.

Some notes regarding "Services changing while the router is being
rebuilt":

* This PR sacrifices some "correctness" for speed: when a service
  is changed, it takes time for that change to be propagated to other
  workers and kong nodes. If a service is changed it will take some
  seconds before that change is made effective in the router.

* Consider that even on the current case (were no caching is used),
  if the Service changes while the router iterates over route number
  501, the first 500 routes will be "wrong" anyway, until the router
  gets refreshed - so at the end we don't loose that much with this
  change.